repairs

+++ PROEProjekt2/car.cpp
@@ -10,13 +10,13 @@ shared_ptr<Vehicle> Vehicle::makeMotorcycle(string mod, double price, int ID, st
 {
     return shared_ptr<Vehicle>(new Motorcycle(mod, price, ID, cond, eng, eng_type));
 }
-double Car::Get_Price() const {return price_;}
+int Car::Get_Price() const {return price_;}
 int Car::Get_ID() const {return ID_;}
 string Car::Get_Condition() const {return condition_;}
 string Car::Get_Engine() const {return engine_;}
 string Car::Get_Model() const {return model_;}
 
-double Motorcycle::Get_Price() const {return price_;}
+int Motorcycle::Get_Price() const {return price_;}
 int Motorcycle::Get_ID() const {return ID_;}
 string Motorcycle::Get_Condition() const {return condition_;}
 string Motorcycle::Get_Engine() const {return engine_;}
