+++ PROEProjekt2/car.h
@@ -15,7 +15,7 @@ class Vehicle
 {
 protected:
     string model_;
-    double price_;
+    int price_;
     string condition_;
     int ID_;
     string engine_;
@@ -23,7 +23,7 @@ public:
     static shared_ptr<Vehicle> makeCar(string, double, int, string, string, string);
     static shared_ptr<Vehicle> makeMotorcycle(string, double, int, string, string, string);
     Vehicle(string mod, double price, int ID, string cond, string eng)  {model_=mod; price_=price; ID_=ID; condition_=cond; engine_=eng;}
-    virtual double Get_Price() const =0;
+    virtual int Get_Price() const =0;
     virtual int Get_ID() const =0;
     virtual string Get_Model() const = 0;
     virtual string Get_Condition() const =0;
@@ -64,7 +64,7 @@ public:
     Car(): Vehicle("BMW 330d", 230000, 1, "Nowy", "Diesel") {transmission_="Automatyczna Hydrauliczna";}
     Car(string mod, double price, int ID, string cond, string eng, string transmission): Vehicle(mod, price, ID, cond, eng) {transmission_=transmission;}
     virtual string Get_Model() const override;
-    virtual double Get_Price() const override;
+    virtual int Get_Price() const override;
     virtual int Get_ID() const override;
     virtual string Get_Condition() const override;
     virtual string Get_Engine() const override;
@@ -96,7 +96,7 @@ public:
     Motorcycle(string mod, double price, int ID, string cond, string eng, string engine_type): Vehicle(mod, price, ID, cond, eng) {engine_type_=engine_type;}
     Motorcycle(): Vehicle("BMW R1250RS", 230000, 1, "Nowy", "Benzyna") {engine_type_="Dwusuwowy";}
     virtual string Get_Model() const override;
-    virtual double Get_Price() const override;
+    virtual int Get_Price() const override;
     virtual int Get_ID() const override;
     virtual string Get_Condition() const override;
     virtual string Get_Engine() const override;
